Conflicts:
	build/patch_check/boost_iostreams_flushing/test.cpp
	src/dose/dose_internal.ss/src/include/Safir/Dob/Internal/TimeRegistration.h

+++ src/lluf/lluf_utils.ss/tests/ProcessMonitor_test/ProcessMonitor_test.cpp
@@ -81,5 +81,5 @@ int main(int argc, char** argv)
 
     ioService.run();
     thread.join();
-    return pids.size();
+    return static_cast<int>(pids.size());
 }
